# specs/035-protected-resource-subresources/contracts/protected-resources.openapi.yaml
...
openapi: 3.0.3
info:
  title: Third-Party Service Protected Resources (delta)
  version: 1.0.0
  description: >
    Manage a third-party OAuth2 service's protected resource set as first-class, individually addressable
    sub-resources (keyed by percent-encoded normalized URI), plus the optimistic-concurrency change to the
    full-service update. Authentication/authorization are identical to the existing admin
    service-management endpoints (FR-014).
paths:
...
  /api/services/{service-id}:
    parameters:
      - $ref: '#/components/parameters/ServiceId'
    put:
      summary: Update a service (protected_resources now optional + optimistic concurrency)
      operationId: updateService
      tags: [Services]
      parameters:
        - name: If-Match
          in: header
          required: false
          description: >
            REQUIRED when the request body includes `protected_resources` (authoritative set replacement).
            Value is the service ETag/version from a prior GET. Omitted while replacing → 428; stale → 412.
            Not needed when `protected_resources` is omitted.
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            # Same schema as today EXCEPT protected_resources is nullable/omittable with replace semantics.
            schema: { $ref: '#/components/schemas/ServiceUpdateRequest' }
      responses:
        '200':
          description: Service updated.
          headers:
            ETag: { schema: { type: string } }
        '400': { $ref: '#/components/responses/InvalidResourceUri' }
        '404': { $ref: '#/components/responses/ServiceNotFound' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
        '412':
          description: If-Match did not match current version — the resource set changed; re-GET and retry.
          content:
            application/json: { schema: { $ref: '#/components/schemas/Error' } }
        '428':
          description: protected_resources present but If-Match header missing (replacement requires it).
          content:
            application/json: { schema: { $ref: '#/components/schemas/Error' } }
        '500': { $ref: '#/components/responses/InternalServerError' }
components:
  parameters:
    ServiceId:
      name: service-id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    ServiceUpdateRequest:
      description: >
        Existing service update body. Only the protected_resources field changes semantics; all other
        fields are as documented on the canonical Service schema.
      type: object
      properties:
        protected_resources:
          type: array
          nullable: true
          description: >
            OPTIONAL. Omit/null → leave the current set unchanged. Present (incl. []) → authoritatively
            REPLACE the set; requires If-Match. Empty array clears the set.
          items: { type: string, format: uri }
      additionalProperties: true
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        message: { type: string }
  responses:
    ServiceNotFound:
      description: No service with the given id (FR-010).
      content:
        application/json: { schema: { $ref: '#/components/schemas/Error' } }
    InvalidResourceUri:
      description: Malformed/relative/empty/whitespace URI — rejected before any state change (FR-006).
      content:
        application/json: { schema: { $ref: '#/components/schemas/Error' } }
    ResourceConflict:
      description: URI already owned by another service, or rename target already present (FR-007).
      content:
        application/json: { schema: { $ref: '#/components/schemas/Error' } }
    InternalServerError:
      description: Unexpected server error.
      content:
        application/json: { schema: { $ref: '#/components/schemas/Error' } }